Update: Creation of liteNode moved to navbar
 - This has been done, so that liteNodes can be created dynamically, with a corresponding name, once a menu button is clicked
