+++ Frontend/src/app/Workspace/workspace-boundary/workspace-boundary.component.css
@@ -10,3 +10,7 @@
 #importFromCommunity{
   display: none;
 }
+
+#workspaceCanvas{
+  position: relative;
+}
