+++ Frontend/src/app/Workspace/workspace-boundary/workspace-boundary.component.ts
@@ -6,7 +6,6 @@ import {Store} from "@ngxs/store";
 import projectList from "../import/import.component";
 import {WorkspaceState} from "../../../Storage/workspace";
 import * as LeaderLine from "leader-line-new";
-import * as litegraph from "litegraph.js";
 
 @Component({
 	selector: 'app-workspace-boundary',
@@ -14,10 +13,6 @@ import * as litegraph from "litegraph.js";
 	styleUrls: ['./workspace-boundary.component.css']
 })
 export class WorkspaceBoundaryComponent implements OnInit {
-
-	liteNodes: litegraph.LGraph[];
-
-	graph: litegraph.LGraph;
 
 
   public projectL = projectList
@@ -31,24 +26,6 @@ export class WorkspaceBoundaryComponent implements OnInit {
 	}
 
 	ngOnInit(): void {
-		this.liteNodes = [];
-
-		this.graph = new litegraph.LGraph();
-
-		let canvas = new litegraph.LGraphCanvas("#mycanvas", this.graph);
-
-		let node_const = litegraph.LiteGraph.createNode("basic/const");
-		node_const.pos = [200,200];
-		this.graph.add(node_const);
-		node_const.setValue(4.5);
-
-		let node_watch = litegraph.LiteGraph.createNode("basic/watch");
-		node_watch.pos = [700,200];
-		this.graph.add(node_watch);
-
-		node_const.connect(0, node_watch, 0 );
-
-		this.graph.start()
 	}
 
 	reload() {
@@ -64,13 +41,5 @@ export class WorkspaceBoundaryComponent implements OnInit {
 		}
 	}
 
-	createLiteNode()
-	{
-		let node_const = litegraph.LiteGraph.createNode("basic/const");
-		node_const.pos = [200,200];
-		this.graph.add(node_const);
-		node_const.setValue(3);
-		node_const.title = "TestHello";
-	}
 
 }
